# swagger-api.yaml
openapi: 3.0.4
swagger: "2.0"
info:
  title: Swagger Warmhouse - OpenAPI 3.0
  version: 1.0.0
servers:
  - url: http://localhost:8084/api/v1
security:
  - XAuthToken: []
...
paths:
...
  /houses:
    get:
      tags: [Houses]
      summary: List houses
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by owner (user_id)
      responses:
        "200":
          description: Houses list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseList"
    post:
      tags: [Houses]
      summary: Create house
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseCreate"
      responses:
        "201":
          description: Created house
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
        "400":
          $ref: "#/components/responses/BadRequest"
  /houses/{houseId}:
    get:
      tags: [Houses]
      summary: Get house by id
      parameters:
        - $ref: "#/components/parameters/HouseId"
      responses:
        "200":
          description: House
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Houses]
      summary: Update house (partial)
      parameters:
        - $ref: "#/components/parameters/HouseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseUpdate"
      responses:
        "200":
          description: Updated house
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Houses]
      summary: Delete house
      parameters:
        - $ref: "#/components/parameters/HouseId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /sensors:
    get:
      tags: [Sensors]
      summary: List sensors
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: house_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by house_id
        - name: type_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by type_id
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by status
      responses:
        "200":
          description: Sensors list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorList"
    post:
      tags: [Sensors]
      summary: Create sensor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorCreate"
      responses:
        "201":
          description: Created sensor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "400":
          $ref: "#/components/responses/BadRequest"
  /sensors/{sensorId}:
    get:
      tags: [Sensors]
      summary: Get sensor by id
      parameters:
        - $ref: "#/components/parameters/SensorId"
      responses:
        "200":
          description: Sensor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Sensors]
      summary: Update sensor (partial)
      parameters:
        - $ref: "#/components/parameters/SensorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorUpdate"
      responses:
        "200":
          description: Updated sensor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Sensors]
      summary: Delete sensor
      parameters:
        - $ref: "#/components/parameters/SensorId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    XAuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: Auth token returned by `/auth/login`. Send it in header `X-Auth-Token`.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    HouseId:
      name: houseId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    SensorId:
      name: sensorId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBadRequest"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorNotFound"
  schemas:
    ErrorBadRequest:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Bad request error"
        details:
          type: object
          additionalProperties: true
      required: [code, message]
    ErrorNotFound:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          example: "NOT FOUND"
        message:
          type: string
          example: "Not found error"
        details:
          type: object
          additionalProperties: true
      required: [code, message]

    # ---------- Users ----------
...
    House:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        address:
          type: string
      required: [id, user_id, address]
    HouseCreate:
      type: object
      additionalProperties: false
      properties:
        user_id:
          type: integer
          format: int64
        address:
          type: string
      required: [user_id, address]
    HouseUpdate:
      type: object
      additionalProperties: false
      properties:
        address:
          type: string
      minProperties: 1
    HouseList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/House"
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      required: [items, limit, offset, total]

    # ---------- Sensors ----------
    Sensor:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        house_id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
        type_id:
          type: integer
          format: int64
        location:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        value:
          type: number
          format: float
          default: 0
        unit:
          type: string
          maxLength: 20
        status:
          type: string
          default: inactive
        last_updated:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      required:
        - id
        - house_id
        - name
        - type_id
        - location
        - serial_number
        - status
        - last_updated
        - created_at
    SensorCreate:
      type: object
      additionalProperties: false
      properties:
        house_id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
        type_id:
          type: integer
          format: int64
        location:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        value:
          type: number
          format: float
          default: 0
        unit:
          type: string
          maxLength: 20
        status:
          type: string
          default: inactive
      required: [house_id, name, type_id, location, serial_number, status]
    SensorUpdate:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 100
        type_id:
          type: integer
          format: int64
        location:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        value:
          type: number
          format: float
        unit:
          type: string
          maxLength: 20
        status:
          type: string
      minProperties: 1
    SensorList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Sensor"
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      required: [items, limit, offset, total]

    # ---------- Devices ----------
...